student_details/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asteval==0.9.31
astunparse==1.6.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
boto3==1.28.69
botocore==1.31.69
Brlapi==0.8.3
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.10
bw2parameters==1.1.0
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constructive-geometries==0.8.2
country-converter==1.0.0
cryptography==3.4.8
cupshelpers==1.0
datapackage==1.15.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
duplicity==0.8.21
eight==1.0.1
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.12.0
flake8==4.0.1
future==0.18.2
greenlet==3.0.0
gyp==0.1
httplib2==0.20.2
idna==3.3
ijson==3.2.3
importlib-metadata==4.6.4
isodate==0.6.1
jeepney==0.7.1
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linear-tsv==1.1.0
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mrio-common-metadata==0.2.1
netifaces==0.11.0
numba==0.58.1
numpy==1.23.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
pandas==2.1.1
paramiko==2.9.3
peewee==3.17.0
pexpect==4.8.0
Pillow==9.0.1
Pint==0.22
platformdirs==3.5.1
prettytable==3.9.0
protobuf==3.12.4
psutil==5.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pyflakes==2.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPrind==2.11.3
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
pyxlsb==1.0.10
PyYAML==5.4.1
referencing==0.30.2
reportlab==3.6.8
requests==2.25.1
rfc3986==2.0.0
rpds-py==0.10.6
s3transfer==0.7.0
scipy==1.11.3
SecretStorage==3.3.1
six==1.16.0
sparse==0.14.0
SQLAlchemy==2.0.22
sqlparse==0.4.2
stats-arrays==0.6.6
systemd-python==234
tableschema==1.20.2
tabulator==1.53.5
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unfold==1.1.5
unicodecsv==0.14.1
Unidecode==1.3.7
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.23.0
voluptuous==0.13.1
wadllib==1.3.6
wcwidth==0.2.8
Whoosh==2.7.4
wrapt==1.15.0
wurst==0.3.4
xdg==5
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.1.9
zipp==1.0.0
